+++ EmpyrionPrime.ModFramework/CHANGELOG.md
@@ -1,4 +1,11 @@
-# Changelog
+# EmpyrionPrime.ModFramework Changelog
+All notable changes to this project will be documented in this file.
+
+The format is loosely based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/).
+
+## Unreleased
+### Removed
+ - EleonExtensions, these can now be found in EmpyrionPrime.Plugin
 
 ## 1.0.1 (2023-04-01)
 ### Fixed