+++ EmpyrionPrime.Plugin/CHANGELOG.md
@@ -1,4 +1,15 @@
-# Changelog
+# EmpyrionPrime.Plugin Changelog
+All notable changes to this project will be documented in this file.
+
+The format is loosely based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/).
+
+## Unreleased
+### Added
+ - Eleon API helper extensions
+   - MessageData to ChatInfo (ToChatInfo)
+   - Int32 to Id
+   - String to PString
+ - Eleon API enumerations: EntityType, FactionGroup, MessagePriority
 
 
 ## 1.0.0 (2023-03-31)